Add files via upload
This is a paper I co-authored that outlines how NLP and ML techniques can be used to predict the ratings a book will get:

Every year millions of people read books and log their interactions on a social cataloging site called GoodReads. GoodReads allows users to review, rank, sort, and explore books, annotations, and quotes in a variety of languages. We took a dataset that is a subset of user provided reviews of young adult novels. These universal, often coming of age, literary works are read by a wide swath of the population all over the world. It is for this reason we decided to build our model from this subset.